Extract other exported interfaces
