+++ src/interfaces/guess_result.ts
@@ -1,0 +1,7 @@
+import KmqMember from "../structures/kmq_member";
+
+export default interface GuessResult {
+    correct: boolean;
+    error?: boolean;
+    correctGuessers?: Array<KmqMember>;
+}
