+++ src/interfaces/success_guess_result.ts
@@ -1,0 +1,5 @@
+export default interface SuccessfulGuessResult {
+    userID: string;
+    pointsEarned: number;
+    expGain: number;
+}
