+++ src/interfaces/unique_song_counter.ts
@@ -1,0 +1,4 @@
+export default interface UniqueSongCounter {
+    uniqueSongsPlayed: number;
+    totalSongs: number;
+}
