+++ src/structures/game_round.ts
@@ -16,10 +16,10 @@ import {
     EMBED_SUCCESS_COLOR,
     getMention,
 } from "../helpers/discord_utils";
-import { UniqueSongCounter } from "./song_selector";
 import { friendlyFormattedNumber } from "../helpers/utils";
 import PlayerRoundResult from "../interfaces/player_round_result";
 import QueriedSong from "../interfaces/queried_song";
+import UniqueSongCounter from "../interfaces/unique_song_counter";
 /** List of characters to remove from song/artist names/guesses */
 // eslint-disable-next-line no-useless-escape
 const REMOVED_CHARACTERS = /[\|’\ '?!.\-,:;★*´\(\)\+\u200B]/g;
@@ -29,7 +29,7 @@ const CHARACTER_REPLACEMENTS = [
     { pattern: /&/g, replacement: "and" },
 ];
 
-export interface GuessCorrectness {
+interface GuessCorrectness {
     exact: boolean;
     similar: boolean;
 }
