+++ src/structures/guild_preference.ts
@@ -46,35 +46,10 @@ import {
 } from "../commands/game_options/answer";
 import Session from "./session";
 import MatchedArtist from "../interfaces/matched_artist";
+import GameOptions from "../interfaces/game_options";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const logger = new IPCLogger("guild_preference");
-
-export interface GameOptions {
-    beginningYear: number;
-    endYear: number;
-    gender: Array<Gender>;
-    limitStart: number;
-    limitEnd: number;
-    seekType: SeekType;
-    specialType: SpecialType;
-    guessModeType: GuessModeType;
-    releaseType: ReleaseType;
-    artistType: ArtistType;
-    answerType: AnswerType;
-    shuffleType: ShuffleType;
-    groups: MatchedArtist[];
-    excludes: MatchedArtist[];
-    includes: MatchedArtist[];
-    goal: number;
-    guessTimeout: number;
-    duration: number;
-    languageType: LanguageType;
-    multiGuessType: MultiGuessType;
-    subunitPreference: SubunitsPreference;
-    ostPreference: OstPreference;
-    forcePlaySongID: string;
-}
 
 type GameOptionValue =
     | number
