+++ src/structures/music_round.ts
@@ -2,9 +2,9 @@ import Eris from "eris";
 import { EMBED_SUCCESS_BONUS_COLOR } from "../helpers/discord_utils";
 import PlayerRoundResult from "../interfaces/player_round_result";
 import QueriedSong from "../interfaces/queried_song";
+import UniqueSongCounter from "../interfaces/unique_song_counter";
 import MessageContext from "./message_context";
 import Round from "./round";
-import { UniqueSongCounter } from "./song_selector";
 
 export default class MusicRound extends Round {
     /** UUID associated with skip interaction custom_id */
