+++ src/structures/song_selector.ts
@@ -16,17 +16,13 @@ import {
 import { chooseWeightedRandom, setDifference } from "../helpers/utils";
 import { ShuffleType } from "../commands/game_options/shuffle";
 import QueriedSong from "../interfaces/queried_song";
+import UniqueSongCounter from "../interfaces/unique_song_counter";
 
 export const LAST_PLAYED_SONG_QUEUE_SIZE = 10;
 export const SELECTION_WEIGHT_VALUES_HARD = [1, 2, 4, 8, 16];
 export const SELECTION_WEIGHT_VALUES_EASY = [
     ...SELECTION_WEIGHT_VALUES_HARD,
 ].reverse();
-
-export interface UniqueSongCounter {
-    uniqueSongsPlayed: number;
-    totalSongs: number;
-}
 
 const logger = new IPCLogger("song_selector");
