+++ src/structures/team_scoreboard.ts
@@ -1,10 +1,8 @@
-import Scoreboard, {
-    SuccessfulGuessResult,
-    SCOREBOARD_FIELD_CUTOFF,
-} from "./scoreboard";
+import Scoreboard, { SCOREBOARD_FIELD_CUTOFF } from "./scoreboard";
 import Player from "./player";
 import Team from "./team";
 import { IPCLogger } from "../logger";
+import SuccessfulGuessResult from "../interfaces/success_guess_result";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const logger = new IPCLogger("team_scoreboard");
